refactor: {domainController}DocsTest로 이름 변경
- package 참조 올바르게 수정
- Docs Test @DisPlayName(" Docs 테스트")로 변경
